+++ packages/assignment-7/src/components/MonthView.tsx
@@ -1,0 +1,122 @@
+import {
+  VStack,
+  Heading,
+  Table,
+  Thead,
+  Tbody,
+  Th,
+  Tr,
+  Td,
+  Text,
+} from '@chakra-ui/react';
+import { getDaysInMonth } from '../lib/utils/date';
+import { formatMonth } from '../lib/utils/dateFormat';
+import { weekDays } from '../constants/constants';
+import { Event } from '../types/types';
+import { ViewBox } from './ViewBox';
+
+interface MonthViewProps {
+  currentDate: Date;
+  holidays: Record<string, string>;
+  filteredEvents: Event[];
+  notifiedEvents: number[];
+}
+
+export const MonthView = ({
+  currentDate,
+  holidays,
+  filteredEvents,
+  notifiedEvents,
+}: MonthViewProps) => {
+  const daysInMonth = getDaysInMonth(
+    currentDate.getFullYear(),
+    currentDate.getMonth()
+  );
+  const firstDayOfMonth = new Date(
+    currentDate.getFullYear(),
+    currentDate.getMonth(),
+    1
+  ).getDay();
+
+  const days = Array.from({ length: daysInMonth }, (_, i) => i + 1);
+  const weeks = [];
+  let week = Array(7).fill(null);
+
+  for (let i = 0; i < firstDayOfMonth; i++) {
+    week[i] = null;
+  }
+
+  for (const day of days) {
+    const dayIndex = (firstDayOfMonth + day - 1) % 7;
+    week[dayIndex] = day;
+    if (dayIndex === 6 || day === daysInMonth) {
+      weeks.push(week);
+      week = Array(7).fill(null);
+    }
+  }
+  return (
+    <VStack data-testid="month-view" align="stretch" w="full" spacing={4}>
+      <Heading size="md">{formatMonth(currentDate)}</Heading>
+      <Table variant="simple" w="full">
+        <Thead>
+          <Tr>
+            {weekDays.map((day) => (
+              <Th key={day} width="14.28%">
+                {day}
+              </Th>
+            ))}
+          </Tr>
+        </Thead>
+        <Tbody>
+          {weeks.map((week, weekIndex) => (
+            <Tr key={weekIndex}>
+              {week.map((day, dayIndex) => {
+                const dateString = day
+                  ? `${currentDate.getFullYear()}-${String(currentDate.getMonth() + 1).padStart(2, '0')}-${String(day).padStart(2, '0')}`
+                  : '';
+                const holiday = holidays[dateString];
+
+                return (
+                  <Td
+                    key={dayIndex}
+                    height="100px"
+                    verticalAlign="top"
+                    width="14.28%"
+                    position="relative"
+                  >
+                    {day && (
+                      <>
+                        <Text fontWeight="bold">{day}</Text>
+                        {holiday && (
+                          <Text color="red.500" fontSize="sm">
+                            {holiday}
+                          </Text>
+                        )}
+                        {filteredEvents
+                          .filter(
+                            (event) => new Date(event.date).getDate() === day
+                          )
+                          .map((event) => {
+                            const isNotified = notifiedEvents.includes(
+                              event.id
+                            );
+                            return (
+                              <ViewBox
+                                key={event.id}
+                                isNotified={isNotified}
+                                title={event.title}
+                              />
+                            );
+                          })}
+                      </>
+                    )}
+                  </Td>
+                );
+              })}
+            </Tr>
+          ))}
+        </Tbody>
+      </Table>
+    </VStack>
+  );
+};
